DEV: Thresholding and ICR
CHANGES: Focussed on Thresholding, the scoring logic for Result was revamped, previously it was (intensity/NEXT_HIGHER_intensity)*100, now it is (intensity/HIGHEST_intensity)*100. Also the Threshold was set to 90. Anything greater than or equal to 90 is considered to be Blank, anything less that is considered not unmarked. TO DO: Rigorous Testing, Try to make the Threshold Dynamic, ICR