CHECKPO: Anchor Detection
Hit the perspective issue (m_transform and warp transform). Currently, anchor detection detects the anchors by finding similar characteistics of the anchor across a given threshold are (near the area wwhere the anchor was present in the annotated image (2% area of the whole image where the anchor is present in the annotated Image), Tried to skew all the images in the TesFolder wrt the annotated Image (if annotated Image is +0.12°, then all the images in the TestFolder will also be exactly +0.12° skewed, but this creates a lot of issues now, hence SKIPPING this skewing for now. Moving to Field Mapping (prev. optionMapping))