PROD: Making the Code Production Ready
CHANGES: Modularised the Code, the Base folder path is now in the config. TODO: TkInter part is yet to me inculded with the main or maybe a separate script. the OMR Sheet Name has to be added. Toggles for the saving of images in fieldMapping and markedOption has to be added. Finally the code needs to be optimised for redundant function calls (there is a redundancy for getting the annotation filename and the classes, will try to make the json pass from one position to other, have to replace the transformed images align with the co-ords, also need to implement the bbox for KEY FIELDs ICR & OMR part).