PROD: Making the Code Production Ready
CHANGES: Modulated the code, Added logging for every module, Added a toggle flag for result visualization in each step. Wrishav is handling the Regular Expression format for the Key Fields and also the Co-ords for the OMR part of the key fields. TODO: Integrate the logging into one log for every run. Add the option to set manual threshold for the Marked Option Consideration in the FrontEnd.